move logic into hellocash class
